requirements.txt
pandas>=1.5.0
numpy>=1.23.0
pyarrow>=12.0.0
scikit-learn>=1.2.0
optuna>=3.2.0
mlflow>=2.4.0
great-expectations>=0.17.0
kagglehub>=0.2.0
lightgbm>=4.0.0
xgboost>=2.0.0
joblib>=1.2.0
pyyaml>=6.0
kaggle
imbalanced-learn